requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
geopy==2.4.0
python-dateutil==2.8.2
Pillow==10.0.0
gunicorn==21.2.0 